chore(mainnav): clean up and file renaming
