+++ src/Mainnav/index.ts
@@ -1,3 +1,3 @@
-export { SgdsMainnav } from './sgds-mainnav-element'
-export type { MainnavExpandSize } from './sgds-mainnav-element'
+export { SgdsMainnav } from './sgds-mainnav'
+export type { MainnavExpandSize } from './sgds-mainnav'
 export { SgdsMainnavItem } from './sgds-mainnav-item'
