+++ stories/MainNav.stories.mdx
@@ -18,7 +18,7 @@ import { partDescription } from "./common";
         category: "props: SgdsMainnav",
         defaultValue: { summary: "lg" },
         type: {
-          summary: "MainNavExpandSize",
+          summary: "MainnavExpandSize",
         },
       },
     },
